+++ dflib/src/main/java/org/dflib/exp/str/StrSplitExp.java
@@ -2,6 +2,7 @@ package org.dflib.exp.str;
 
 import org.dflib.Exp;
 import org.dflib.Series;
+import org.dflib.builder.ValueCompactor;
 import org.dflib.exp.map.MapExp1;
 
 import java.util.function.Function;
@@ -16,17 +17,35 @@ public class StrSplitExp extends MapExp1<String, String[]> {
 
     public static StrSplitExp splitOnChar(Exp<String> exp, char delimiter, int limit) {
         String sDelimiter = String.valueOf(delimiter);
-        return new StrSplitExp(exp, valToSeries(s -> s.split(sDelimiter, limit)));
+        return new StrSplitExp(exp, valToCompactSeries(s -> s.split(sDelimiter, limit)));
     }
 
     public static StrSplitExp splitOnRegex(Exp<String> exp, String regex) {
         Pattern p = Pattern.compile(regex);
-        return new StrSplitExp(exp, valToSeries(s -> p.split(s)));
+        return new StrSplitExp(exp, valToCompactSeries(p::split));
     }
 
     public static StrSplitExp splitOnRegex(Exp<String> exp, String regex, int limit) {
         Pattern p = Pattern.compile(regex);
-        return new StrSplitExp(exp, valToSeries(s -> p.split(s, limit)));
+        return new StrSplitExp(exp, valToCompactSeries(s -> p.split(s, limit)));
+    }
+
+    private static Function<Series<String>, Series<String[]>> valToCompactSeries(Function<String, String[]> op) {
+        ValueCompactor<String> compactor = new ValueCompactor<>();
+        return s -> s.map(v -> {
+
+            if (v != null) {
+                String[] a = op.apply(v);
+                int len = a.length;
+                for (int i = 0; i < len; i++) {
+                    a[i] = compactor.get(a[i]);
+                }
+
+                return a;
+            }
+
+            return null;
+        });
     }
 
     protected StrSplitExp(Exp<String> exp, Function<Series<String>, Series<String[]>> op) {